Former-commit-id: 19791e6a9175863b8fff3491c4c87e03db910785

+++ applications/programs/mpi_classify_CL2D/mpi_classify_CL2D.h
@@ -289,7 +289,7 @@ public:
     GaussianInterpolator gaussianInterpolator;
 
     // Image dimensions
-    int Ydim, Xdim;
+    size_t Ydim, Xdim;
 
     /// Mask for the background
 	MultidimArray<int> mask;
